+++ packages/devextreme/js/__internal/scheduler/constants.ts
@@ -2,6 +2,7 @@ export const LIST_ITEM_DATA_KEY = 'dxListItemData';
 export const LIST_ITEM_CLASS = 'dx-list-item';
 export const APPOINTMENT_SETTINGS_KEY = 'dxAppointmentSettings';
 
+export const VERTICAL_GROUP_ORIENTATION = 'vertical';
 export const HORIZONTAL_GROUP_ORIENTATION = 'horizontal';
 
 export const VIEWS = {
@@ -15,3 +16,10 @@ export const VIEWS = {
   TIMELINE_MONTH: 'timelineMonth',
   AGENDA: 'agenda',
 };
+export const VIEW_TYPES = Object.values(VIEWS);
+export const TIMELINE_VIEWS = [
+  VIEWS.TIMELINE_DAY,
+  VIEWS.TIMELINE_WEEK,
+  VIEWS.TIMELINE_WORK_WEEK,
+  VIEWS.TIMELINE_MONTH,
+];
