+++ packages/devextreme/js/__internal/scheduler/m_appointment_drag_behavior.ts
@@ -3,7 +3,7 @@ import { Deferred } from '@js/core/utils/deferred';
 import { extend } from '@js/core/utils/extend';
 import Draggable from '@js/ui/draggable';
 
-import { APPOINTMENT_SETTINGS_KEY, LIST_ITEM_DATA_KEY } from './m_constants';
+import { APPOINTMENT_SETTINGS_KEY, LIST_ITEM_DATA_KEY } from './constants';
 import { isSchedulerComponent } from './utils/is_scheduler_component';
 
 const APPOINTMENT_ITEM_CLASS = 'dx-scheduler-appointment';
