+++ packages/devextreme/js/__internal/scheduler/r1/utils/__tests__/base.test.ts
@@ -1,7 +1,7 @@
 import {
   describe, expect, it, jest,
 } from '@jest/globals';
-import { HORIZONTAL_GROUP_ORIENTATION, VERTICAL_GROUP_ORIENTATION } from '@ts/scheduler/r1/const';
+import { HORIZONTAL_GROUP_ORIENTATION, VERTICAL_GROUP_ORIENTATION } from '@ts/scheduler/constants';
 
 import type { ViewType } from '../../types';
 import {
