+++ packages/devextreme/js/__internal/scheduler/r1/utils/views.ts
@@ -1,6 +1,6 @@
 import { isObject } from '@js/core/utils/type';
 
-import { VIEW_TYPES } from '../const';
+import { VIEW_TYPES } from '../../constants';
 import type { ViewType } from '../types';
 
 export const getCurrentView = (
