+++ packages/devextreme/js/__internal/scheduler/workspaces/m_agenda.ts
@@ -10,6 +10,7 @@ import { each } from '@js/core/utils/iterator';
 import { setHeight, setOuterHeight } from '@js/core/utils/size';
 import { agendaUtils, formatWeekday, getVerticalGroupCountClass } from '@ts/scheduler/r1/utils/index';
 
+import { VIEWS } from '../constants';
 import {
   DATE_TABLE_CLASS,
   DATE_TABLE_ROW_CLASS,
@@ -17,7 +18,6 @@ import {
   GROUP_ROW_CLASS,
   TIME_PANEL_CLASS,
 } from '../m_classes';
-import { VIEWS } from '../m_constants';
 import tableCreatorModule from '../m_table_creator';
 import { createReducedResourcesTree, getDataAccessors, getPathToLeaf } from '../resources/m_utils';
 import WorkSpace from './m_work_space';
