+++ packages/devextreme/js/__internal/scheduler/workspaces/m_timeline_day.ts
@@ -1,6 +1,6 @@
 import registerComponent from '@js/core/component_registrator';
 
-import { VIEWS } from '../m_constants';
+import { VIEWS } from '../constants';
 import SchedulerTimeline from './m_timeline';
 
 const TIMELINE_CLASS = 'dx-scheduler-timeline-day';
