+++ packages/devextreme/js/__internal/scheduler/workspaces/m_timeline_month.ts
@@ -4,7 +4,7 @@ import dateUtils from '@js/core/utils/date';
 import { HeaderPanelComponent } from '@ts/scheduler/r1/components/index';
 import { formatWeekdayAndDay, monthUtils } from '@ts/scheduler/r1/utils/index';
 
-import { VIEWS } from '../m_constants';
+import { VIEWS } from '../constants';
 import SchedulerTimeline from './m_timeline';
 
 const TIMELINE_CLASS = 'dx-scheduler-timeline-month';
