+++ packages/devextreme/js/__internal/scheduler/workspaces/m_timeline_work_week.ts
@@ -3,7 +3,7 @@ import {
   getWeekendsCount,
 } from '@ts/scheduler/r1/utils/index';
 
-import { VIEWS } from '../m_constants';
+import { VIEWS } from '../constants';
 import SchedulerTimelineWeek from './m_timeline_week';
 
 const TIMELINE_CLASS = 'dx-scheduler-timeline-work-week';
