+++ packages/devextreme/js/__internal/scheduler/workspaces/m_work_space_day.ts
@@ -1,6 +1,6 @@
 import registerComponent from '@js/core/component_registrator';
 
-import { VIEWS } from '../m_constants';
+import { VIEWS } from '../constants';
 import SchedulerWorkSpaceVertical from './m_work_space_vertical';
 
 const DAY_CLASS = 'dx-scheduler-work-space-day';
