+++ packages/devextreme/js/__internal/scheduler/workspaces/m_work_space_month.ts
@@ -7,7 +7,7 @@ import { hasWindow } from '@js/core/utils/window';
 import { DateTableMonthComponent } from '@ts/scheduler/r1/components/index';
 import { formatWeekday, monthUtils } from '@ts/scheduler/r1/utils/index';
 
-import { VIEWS } from '../m_constants';
+import { VIEWS } from '../constants';
 import { utils } from '../m_utils';
 import SchedulerWorkSpace from './m_work_space_indicator';
