+++ packages/devextreme/js/__internal/scheduler/workspaces/m_work_space_week.ts
@@ -1,7 +1,7 @@
 import registerComponent from '@js/core/component_registrator';
 import { weekUtils } from '@ts/scheduler/r1/utils/index';
 
-import { VIEWS } from '../m_constants';
+import { VIEWS } from '../constants';
 import SchedulerWorkSpaceVertical from './m_work_space_vertical';
 
 const WEEK_CLASS = 'dx-scheduler-work-space-week';
