+++ packages/devextreme/js/__internal/scheduler/workspaces/view_model/m_date_header_data_generator.ts
@@ -1,5 +1,5 @@
 import dateUtils from '@js/core/utils/date';
-import { VIEWS } from '@ts/scheduler/m_constants';
+import { VIEWS } from '@ts/scheduler/constants';
 import {
   formatWeekdayAndDay, getDisplayedCellCount, getGroupCount,
   getHeaderCellText,
